# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-desk/summary:
    get:
      operationId: getServiceDeskSummary
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Summary, content: {application/json: {schema: {$ref: "#/components/schemas/ServiceDeskSummary"}}}}}
  /service-desk/trends:
    get:
      operationId: getServiceDeskTrends
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Trends, content: {application/json: {schema: {$ref: "#/components/schemas/ServiceDeskTrends"}}}}}
  /service-desk/top-assets:
    get:
      operationId: getServiceDeskTopAssets
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Top assets, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ServiceDeskTopAsset"}}}}}}
  /service-desk/upcoming-jobs:
    get:
      operationId: getServiceDeskUpcomingJobs
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Upcoming jobs, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ServiceDeskUpcomingJob"}}}}}}
  /service-desk/rental-pins:
    get:
      operationId: getServiceDeskRentalPins
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Rental pins, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ServiceDeskRentalPin"}}}}}}
  /service-desk/map:
    get:
      operationId: getServiceDeskMap
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Map pins, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ServiceDeskRentalPin"}}}}}}
  /service-desk/ingestion-log:
    get:
      operationId: getServiceDeskIngestionLog
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Ingestion history, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ServiceDeskIngestionLog"}}}}}}
  /service-desk/dashboard/asset-history:
    get:
      operationId: getServiceDeskAssetHistory
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      parameters: [{name: assetRef, in: query, required: true, schema: {type: string}}]
      responses: {"200": {description: Asset history, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ServiceDeskAssetHistory"}}}}}}
  /service-desk/upload-csv:
    post:
      operationId: uploadServiceDeskCsv
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      parameters: [{name: filename, in: query, required: true, schema: {type: string}}, {name: kind, in: query, required: false, schema: {type: string, enum: [completed, outstanding], default: completed}}]
      requestBody: {required: true, content: {text/csv: {schema: {type: string}}}}
      responses: {"200": {description: Imported, content: {application/json: {schema: {$ref: "#/components/schemas/ServiceDeskUploadResult"}}}}}
  /service-desk/import-sqlite:
    post:
      operationId: importServiceDeskSqlite
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: {type: string, format: binary}
      responses:
        "200":
          description: Historical Service Desk import result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ServiceDeskLegacyImportResult"}
        "400":
          description: Invalid Service Desk database
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /service-desk/sync:
    post:
      operationId: syncServiceDesk
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: Sync result, content: {application/json: {schema: {$ref: "#/components/schemas/ServiceDeskSyncResult"}}}}}
  /service-desk/hubspot/properties:
    get:
      operationId: getServiceDeskHubSpotProperties
      tags: [service-desk]
      security: [{ bearerAuth: [] }]
      responses: {"200": {description: HubSpot properties, content: {application/json: {schema: {type: object, additionalProperties: true}}}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServiceDeskLegacyImportResult:
      type: object
      required: [alreadyImported, counts, mapReady]
      properties:
        alreadyImported: {type: boolean}
        mapReady: {type: integer}
        counts:
          type: object
          required: [tickets, jobs, assets, sites, postcodes, ingestionLogs, processedFiles]
          properties:
            tickets: {type: integer}
            jobs: {type: integer}
            assets: {type: integer}
            sites: {type: integer}
            postcodes: {type: integer}
            ingestionLogs: {type: integer}
            processedFiles: {type: integer}
    ServiceDeskSummary:
      type: object
      additionalProperties: true
      required: [openTicketsNow, ticketsThisMonth, jobsThisMonth, activeDaysThisMonth, lastSyncedAt, lastSyncStatus, lastSyncError, outstandingJobsLastUpdatedAt]
      properties:
        openTicketsNow: {type: integer}
        ticketsThisMonth: {type: integer}
        jobsThisMonth: {type: integer}
        activeDaysThisMonth: {type: integer}
        lastSyncedAt: {type: [string, "null"]}
        lastSyncStatus: {type: string}
        lastSyncError: {type: [string, "null"]}
        outstandingJobsLastUpdatedAt: {type: [string, "null"]}
    ServiceDeskTrends:
      type: object
      required: [daily, monthly, period]
      properties:
        daily: {type: array, items: {type: object, additionalProperties: true}}
        monthly: {type: array, items: {type: object, additionalProperties: true}}
        period: {type: integer}
    ServiceDeskTopAsset:
      type: object
      additionalProperties: true
      required: [assetId, assetRef, ticketCount, jobCount, totalCount, totalJobCount, lastServiceDate, installationDate]
      properties:
        assetId: {type: string}
        assetRef: {type: string}
        ticketCount: {type: integer}
        jobCount: {type: integer}
        totalCount: {type: integer}
        totalJobCount: {type: integer}
        lastServiceDate: {type: [string, "null"]}
        installationDate: {type: [string, "null"]}
    ServiceDeskUpcomingJob:
      type: object
      additionalProperties: true
      properties:
        jobRef: {type: [string, "null"]}
        jobType: {type: [string, "null"]}
        site: {type: [string, "null"]}
        assetRef: {type: [string, "null"]}
        engineer: {type: [string, "null"]}
        status: {type: [string, "null"]}
        scheduledDate: {type: string}
    ServiceDeskRentalPin:
      type: object
      required: [assetRef, description, customer, site, postcode, lat, lng, lastServicedDate, serviceType]
      properties:
        assetRef: {type: [string, "null"]}
        description: {type: [string, "null"]}
        customer: {type: [string, "null"]}
        site: {type: [string, "null"]}
        postcode: {type: [string, "null"]}
        lat: {type: number}
        lng: {type: number}
        lastServicedDate: {type: [string, "null"]}
        serviceType: {type: [string, "null"]}
    ServiceDeskIngestionLog:
      type: object
      additionalProperties: true
      required: [id, timestamp, source, status, filesFound, filesProcessed, ticketsImported, errorMessage]
      properties:
        id: {type: integer}
        timestamp: {type: string}
        source: {type: string}
        status: {type: string}
        filesFound: {type: [integer, "null"]}
        filesProcessed: {type: [integer, "null"]}
        ticketsImported: {type: [integer, "null"]}
        errorMessage: {type: [string, "null"]}
    ServiceDeskAssetHistory:
      type: object
      required: [jobRef, jobType, site, engineer, completionDate]
      properties:
        jobRef: {type: string}
        jobType: {type: [string, "null"]}
        site: {type: [string, "null"]}
        engineer: {type: [string, "null"]}
        completionDate: {type: string}
    ServiceDeskUploadResult:
      type: object
      required: [filename, rowsImported]
      properties:
        filename: {type: string}
        rowsImported: {type: integer}
    ServiceDeskSourceResult:
      type: object
      required: [success, count, error]
      properties:
        success: {type: boolean}
        count: {type: integer}
        error: {type: [string, "null"]}
    ServiceDeskSyncResult:
      type: object
      required: [success, message, hubspot, ftps]
      properties:
        success: {type: boolean}
        message: {type: string}
        hubspot: {$ref: "#/components/schemas/ServiceDeskSourceResult"}
        ftps: {$ref: "#/components/schemas/ServiceDeskSourceResult"}
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error